Changed macroses to bool type, rewrote pow func.

+++ project/include/utils.h
@@ -6,8 +6,8 @@
 #define PROJECT_INCLUDE_UTILS_H_
 
 
-size_t timer_from(int from);
+size_t timer_from(unsigned int from);
 
-int custom_pow(int base, int power);
+float custom_pow(int base, int power);
 
 #endif  // PROJECT_INCLUDE_UTILS_H_
